refactor: arrange next directries
