+++ next/src/app/page.tsx
@@ -1,96 +1,68 @@
-import Image from 'next/image'
-import styles from './page.module.css'
+'use client'
 
-export default function Home() {
+import {
+  Heading,
+  Input,
+  Text,
+  HStack,
+  Flex,
+  VStack,
+  Box,
+  Stack,
+  Image,
+  Button,
+} from '@chakra-ui/react'
+import type { NextPage } from 'next'
+import NextLink from 'next/link'
+import TextIconLink from '@/components/molecules/TextIconLink'
+
+const Top: NextPage = () => {
   return (
-    <div className={styles.page}>
-      <main className={styles.main}>
+    <>
+      <Box my={10}>
+        <Stack spacing={8}>
+          <Heading>お土産との出会いを見つけるサービス</Heading>
+          <VStack spacing={6}>
+            <Text>まずはあなたにおすすめのお土産をいくつかご紹介します。</Text>
+            <HStack>
+              <Input placeholder="都市名" size="md" />
+              <Input placeholder="カテゴリー" size="md" />
+            </HStack>
+            <Button variant="primary" as={NextLink} href="./recommend">
+              おすすめのお土産をみる
+            </Button>
+          </VStack>
+          <Flex justifyContent="right">
+            <TextIconLink
+              iconName="FaChevronRight"
+              iconPosition="right"
+              href="./sign_in"
+            >
+              ログインはこちらから
+            </TextIconLink>
+          </Flex>
+        </Stack>
+      </Box>
+      <Box
+        position="fixed"
+        top={0}
+        left={0}
+        width="100vw"
+        height="100dvh"
+        zIndex={-1}
+      >
         <Image
-          className={styles.logo}
-          src="https://nextjs.org/icons/next.svg"
-          alt="Next.js logo"
-          width={180}
-          height={38}
-          priority
+          src="/images/bg_top.jpg"
+          alt="トップページ"
+          w="100%"
+          h="100%"
+          objectFit="cover"
+          objectPosition="center"
+          opacity="30%"
         />
-        <ol>
-          <li>
-            Get started by editing <code>src/app/page.tsx</code>.
-          </li>
-          <li>Save and see your changes instantly.</li>
-          <li>test deploy again and again</li>
-        </ol>
-
-        <div className={styles.ctas}>
-          <a
-            className={styles.primary}
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              className={styles.logo}
-              src="https://nextjs.org/icons/vercel.svg"
-              alt="Vercel logomark"
-              width={20}
-              height={20}
-            />
-            Deploy now
-          </a>
-          <a
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-            className={styles.secondary}
-          >
-            Read our docs
-          </a>
-        </div>
-      </main>
-      <footer className={styles.footer}>
-        <a
-          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="https://nextjs.org/icons/file.svg"
-            alt="File icon"
-            width={16}
-            height={16}
-          />
-          Learn
-        </a>
-        <a
-          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="https://nextjs.org/icons/window.svg"
-            alt="Window icon"
-            width={16}
-            height={16}
-          />
-          Examples
-        </a>
-        <a
-          href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="https://nextjs.org/icons/globe.svg"
-            alt="Globe icon"
-            width={16}
-            height={16}
-          />
-          Go to nextjs.org →
-        </a>
-      </footer>
-    </div>
+      </Box>
+    </>
   )
 }
+
+export default Top
