+++ next/src/app/sign_in/page.tsx
@@ -1,7 +1,7 @@
 import { Heading, HStack, Spacer } from '@chakra-ui/react'
 import type { NextPage } from 'next'
-import TextIconLink from '../components/molecules/TextIconLink'
-import SigninForm from '../components/organisms/form/SigninForm'
+import TextIconLink from '@/components/molecules/TextIconLink'
+import SigninForm from '@/components/organisms/form/SigninForm'
 
 const Login: NextPage = () => {
   return (
