+++ next/src/components/molecules/DataWithIcon.tsx
@@ -1,7 +1,7 @@
 import { HStack } from '@chakra-ui/react'
 import { ReactNode } from 'react'
 import CustomIcon from '../atoms/CustomIcon'
-import { iconMapper } from '@/app/iconMapper'
+import { iconMapper } from '@/utils/iconMapper'
 
 type DataWithIconProps = {
   iconName: keyof typeof iconMapper
