+++ next/src/providers/providers.tsx
@@ -1,6 +1,6 @@
 'use client'
 import { ChakraProvider } from '@chakra-ui/react'
-import customTheme from './theme'
+import customTheme from '@/styles/theme'
 
 export function Providers({ children }: { children: React.ReactNode }) {
   return <ChakraProvider theme={customTheme}>{children}</ChakraProvider>
