+++ next/src/utils/iconMapper.ts
